feat: add XHTTP/SplitHTTP transport support and Config-based timeout configuration
- Add full XHTTP/SplitHTTP transport implementation ported from mihomo's
  transport/xhttp/, supporting all three modes: stream-one, stream-up,
  packet-up with header padding, XMUX reuse, gRPC header camouflage,
  H2/HTTP1.1/H3 (QUIC) transport, REALITY TLS, and downloadSettings
  (upstream/downstream separation)
- XHTTP works with VLESS, VMess, and Trojan protocols
- Add Config struct with configurable DialTimeout, DirectTimeout, and
  XHTTPDialTimeout, replacing hardcoded timeout values throughout
- Update FromURL/FromURLs API to accept Config instead of time.Duration
- Preserve backward-compatible Direct singleton
- Refactor across all files: extract shared helpers (extractVMessPayload,
  decodeVMessJSON, resolveProxyIP, anyToString, extractUserPass,
  tryParseSSJSON, awaitDownload, makeTransportMaker, buildXPaddingConfig),
  eliminate duplicated code, simplify control flow
- Add 95 XHTTP tests covering config parsing, range/session ID/padding
  generation, conn wrapper, wait-read-closer, reuse manager, transport
  params, TLS/REALITY/H3, downloadSettings, proxy creation for all
  protocols, and factory integration
- Fix bugs: empty range fallback now returns {0,0} instead of erroring,
  TLS config no longer created when no TLS params are present
- Update README with Config documentation and new API examples
- 360 total tests passing with -race, gofmt and vet clean

+++ helpers_test.go
@@ -25,7 +25,7 @@ func TestParsePort(t *testing.T) {
 		{"int64", int64(80), 80, false},
 		{"uint16", uint16(443), 443, false},
 		{"float64 whole", float64(8080), 8080, false},
-		{"float64 fractional", float64(8080.5), 0, true},
+		{"float64 fractional", 8080.5, 0, true},
 		{"nil", nil, 0, true},
 		{"json.Number", json.Number("9090"), 9090, false},
 	}
